requirements.txt
# Streamlit Dashboard Dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
plotly==5.18.0
pyyaml==6.0.1
python-dotenv==1.0.0
